[WIP] StaticMeshGenerator: Keep order of materials in slots
For some reason, the order of materials in a mesh's material slots
sometimes gets shuffled for no reason. This is easy to fix, but is
a pain in the butt to do manually.

Instead of blindly assigning materials to slots, take into account
slot names and use this information to rebuild the original order.
Note that materials from LODs are also included, but since they're
not used (as the asset generator does not generate LODs), the logs
can and will have some warnings about some materials.

Marked as WIP because I have not fully checked whether the results
are fully sane. However, first tests are very promising: no issues
with order of materials in their slots, and no missing materials.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>